requirements.txt
python-telegram-bot==13.13
psycopg2-binary==2.9.9
pandas==2.1.4
loguru==0.7.2
sqlalchemy==2.0.30
python-dotenv==1.0.1
matplotlib==3.9.0
openai==1.60.2